+++ src/OpenTelemetry/Trace/SamplingResult.cs
@@ -13,6 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
+
 using System.Collections.Generic;
 using System.Linq;
 
@@ -26,22 +27,32 @@ namespace OpenTelemetry.Trace
         /// <summary>
         /// Initializes a new instance of the <see cref="SamplingResult"/> struct.
         /// </summary>
-        /// <param name="isSampled">True if sampled, false otherwise.</param>
-        public SamplingResult(bool isSampled)
+        /// <param name="decision"> indicates whether an activity object is recorded and sampled.</param>
+        public SamplingResult(SamplingDecision decision)
         {
-            this.IsSampled = isSampled;
+            this.Decision = decision;
             this.Attributes = Enumerable.Empty<KeyValuePair<string, object>>();
         }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="SamplingResult"/> struct.
         /// </summary>
-        /// <param name="isSampled">True if sampled, false otherwise.</param>
+        /// <param name="isSampled"> True if sampled, false otherwise.</param>
+        public SamplingResult(bool isSampled)
+        {
+            this.Decision = isSampled ? SamplingDecision.RecordAndSampled : SamplingDecision.NotRecord;
+            this.Attributes = Enumerable.Empty<KeyValuePair<string, object>>();
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="SamplingResult"/> struct.
+        /// </summary>
+        /// <param name="decision">indicates whether an activity object is recorded and sampled.</param>
         /// <param name="attributes">Attributes associated with the sampling decision. Attributes list passed to
         /// this method must be immutable. Mutations of the collection and/or attribute values may lead to unexpected behavior.</param>
-        public SamplingResult(bool isSampled, IEnumerable<KeyValuePair<string, object>> attributes)
+        public SamplingResult(SamplingDecision decision, IEnumerable<KeyValuePair<string, object>> attributes)
         {
-            this.IsSampled = isSampled;
+            this.Decision = decision;
 
             // Note: Decision object takes ownership of the collection.
             // Current implementation has no means to ensure the collection will not be modified by the caller.
@@ -50,10 +61,9 @@ namespace OpenTelemetry.Trace
         }
 
         /// <summary>
-        /// Gets a value indicating whether Span was sampled or not.
-        /// The value is not suppose to change over time and can be cached.
+        /// Gets a value indicating indicates whether an activity object is recorded and sampled.
         /// </summary>
-        public bool IsSampled { get; }
+        public SamplingDecision Decision { get; }
 
         /// <summary>
         /// Gets a map of attributes associated with the sampling decision.
@@ -83,14 +93,14 @@ namespace OpenTelemetry.Trace
             }
 
             var that = (SamplingResult)obj;
-            return this.IsSampled == that.IsSampled && this.Attributes == that.Attributes;
+            return this.Decision == that.Decision && this.Attributes == that.Attributes;
         }
 
         /// <inheritdoc/>
         public override int GetHashCode()
         {
             var result = 1;
-            result = (31 * result) + this.IsSampled.GetHashCode();
+            result = (31 * result) + this.Decision.GetHashCode();
             result = (31 * result) + this.Attributes.GetHashCode();
             return result;
         }