+++ test/Benchmarks/OpenTelemetrySdkBenchmarks.cs
@@ -18,7 +18,6 @@ using BenchmarkDotNet.Attributes;
 using Benchmarks.Tracing;
 using OpenTelemetry;
 using OpenTelemetry.Trace;
-using OpenTelemetry.Trace.Samplers;
 
 namespace Benchmarks
 {