+++ test/OpenTelemetry.Tests/Implementation/Trace/Export/SimpleActivityProcessorTest.cs
@@ -19,7 +19,7 @@ using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
 using OpenTelemetry.Testing.Export;
-using OpenTelemetry.Trace.Samplers;
+using OpenTelemetry.Tests.Implementation.Trace;
 using Xunit;
 
 namespace OpenTelemetry.Trace.Test
@@ -30,22 +30,6 @@ namespace OpenTelemetry.Trace.Test
         private const string SpanName2 = "MySpanName/2";
         private const string ActivitySourceName = "defaultactivitysource";
 
-        private TestActivityExporter activityExporter;
-        private TracerProvider openTelemetry;
-        private ActivitySource activitySource;
-
-        public SimpleActivityProcessorTest()
-        {
-            this.activityExporter = new TestActivityExporter(null);
-            this.openTelemetry = Sdk.CreateTracerProviderBuilder()
-                .AddSource(ActivitySourceName)
-                .SetSampler(new AlwaysOnSampler())
-                .AddProcessor(new SimpleActivityProcessor(this.activityExporter))
-                .Build();
-
-            this.activitySource = new ActivitySource(ActivitySourceName);
-        }
-
         [Fact]
         public void ThrowsOnNullExporter()
         {
@@ -55,11 +39,11 @@ namespace OpenTelemetry.Trace.Test
         [Fact]
         public void ThrowsInExporter()
         {
-            this.activityExporter = new TestActivityExporter(_ => throw new ArgumentException("123"));
-            this.openTelemetry = Sdk.CreateTracerProviderBuilder()
+            var activityExporter = new TestActivityExporter(_ => throw new ArgumentException("123"));
+            using var openTelemetry = Sdk.CreateTracerProviderBuilder()
                         .AddSource("random")
                         .SetSampler(new AlwaysOnSampler())
-                        .AddProcessor(new SimpleActivityProcessor(this.activityExporter))
+                        .AddProcessor(new SimpleActivityProcessor(activityExporter))
                         .Build();
 
             ActivitySource source = new ActivitySource("random");
@@ -72,10 +56,10 @@ namespace OpenTelemetry.Trace.Test
         [Fact]
         public void ProcessorDoesNotBlockOnExporter()
         {
-            this.activityExporter = new TestActivityExporter(async _ => await Task.Delay(500));
-            this.openTelemetry = Sdk.CreateTracerProviderBuilder()
+            var activityExporter = new TestActivityExporter(async _ => await Task.Delay(500));
+            using var openTelemetry = Sdk.CreateTracerProviderBuilder()
                         .AddSource("random")
-                        .AddProcessor(new SimpleActivityProcessor(this.activityExporter))
+                        .AddProcessor(new SimpleActivityProcessor(activityExporter))
                         .Build();
 
             ActivitySource source = new ActivitySource("random");
@@ -88,8 +72,64 @@ namespace OpenTelemetry.Trace.Test
 
             Assert.InRange(sw.Elapsed, TimeSpan.Zero, TimeSpan.FromMilliseconds(100));
 
-            var exported = this.WaitForSpans(this.activityExporter, 1, TimeSpan.FromMilliseconds(600));
+            var exported = this.WaitForSpans(activityExporter, 1, TimeSpan.FromMilliseconds(600));
 
+            Assert.Single(exported);
+        }
+
+        [Fact]
+        public void ProcessorDoesNotSendRecordDecisionSpanToExporter()
+        {
+            var testSampler = new TestSampler();
+            testSampler.SamplingAction = (samplingParameters) =>
+            {
+                return new SamplingResult(SamplingDecision.Record);
+            };
+
+            using var exporter = new TestActivityExporter(null);
+            using var openTelemetry = Sdk.CreateTracerProviderBuilder()
+                        .AddSource("random")
+                        .AddProcessor(new SimpleActivityProcessor(exporter))
+                        .SetSampler(testSampler)
+                        .Build();
+
+            using ActivitySource source = new ActivitySource("random");
+            var activity = source.StartActivity("somename");
+            activity.Stop();
+
+            Assert.True(activity.IsAllDataRequested);
+            Assert.Equal(ActivityTraceFlags.None, activity.ActivityTraceFlags);
+            Assert.False(activity.Recorded);
+
+            var exported = this.WaitForSpans(exporter, 0, TimeSpan.FromMilliseconds(100));
+            Assert.Empty(exported);
+        }
+
+        [Fact]
+        public void ProcessorSendsRecordAndSampledDecisionSpanToExporter()
+        {
+            var testSampler = new TestSampler();
+            testSampler.SamplingAction = (samplingParameters) =>
+            {
+                return new SamplingResult(SamplingDecision.RecordAndSampled);
+            };
+
+            using var exporter = new TestActivityExporter(null);
+            using var openTelemetry = Sdk.CreateTracerProviderBuilder()
+                        .AddSource("random")
+                        .AddProcessor(new SimpleActivityProcessor(exporter))
+                        .SetSampler(testSampler)
+                        .Build();
+
+            using ActivitySource source = new ActivitySource("random");
+            var activity = source.StartActivity("somename");
+            activity.Stop();
+
+            Assert.True(activity.IsAllDataRequested);
+            Assert.Equal(ActivityTraceFlags.Recorded, activity.ActivityTraceFlags);
+            Assert.True(activity.Recorded);
+
+            var exported = this.WaitForSpans(exporter, 1, TimeSpan.FromMilliseconds(100));
             Assert.Single(exported);
         }
 
@@ -118,10 +158,16 @@ namespace OpenTelemetry.Trace.Test
         [Fact]
         public void ExportDifferentSampledSpans()
         {
+            using var exporter = new TestActivityExporter(null);
+            using var openTelemetrySdk = Sdk.CreateTracerProviderBuilder()
+                                            .AddSource(ActivitySourceName)
+                                            .AddProcessor(new SimpleActivityProcessor(exporter))
+                                            .Build();
+
             var span1 = this.CreateSampledEndedSpan(SpanName1);
             var span2 = this.CreateSampledEndedSpan(SpanName2);
 
-            var exported = this.WaitForSpans(this.activityExporter, 2, TimeSpan.FromMilliseconds(100));
+            var exported = this.WaitForSpans(exporter, 2, TimeSpan.FromMilliseconds(100));
             Assert.Equal(2, exported.Length);
             Assert.Contains(span1, exported);
             Assert.Contains(span2, exported);
@@ -130,6 +176,12 @@ namespace OpenTelemetry.Trace.Test
         [Fact(Skip = "Reenable once AlwaysParentSampler is added")]
         public void ExportNotSampledSpans()
         {
+            using var exporter = new TestActivityExporter(null);
+            using var openTelemetrySdk = Sdk.CreateTracerProviderBuilder()
+                                            .AddSource(ActivitySourceName)
+                                            .AddProcessor(new SimpleActivityProcessor(exporter))
+                                            .Build();
+
             var span1 = this.CreateNotSampledEndedSpan(SpanName1);
             var span2 = this.CreateSampledEndedSpan(SpanName2);
 
@@ -138,7 +190,7 @@ namespace OpenTelemetry.Trace.Test
             // and checking that the first exported span is the sampled span (the non sampled did not get
             // exported).
 
-            var exported = this.WaitForSpans(this.activityExporter, 1, TimeSpan.FromMilliseconds(100));
+            var exported = this.WaitForSpans(exporter, 1, TimeSpan.FromMilliseconds(100));
 
             // Need to check this because otherwise the variable span1 is unused, other option is to not
             // have a span1 variable.
@@ -148,7 +200,6 @@ namespace OpenTelemetry.Trace.Test
 
         public void Dispose()
         {
-            this.activityExporter.ShutdownAsync(CancellationToken.None);
             Activity.Current = null;
         }
 
@@ -156,7 +207,8 @@ namespace OpenTelemetry.Trace.Test
         {
             var context = new ActivityContext(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.Recorded);
 
-            var activity = this.activitySource.StartActivity(spanName, ActivityKind.Internal, context);
+            using ActivitySource activitySource = new ActivitySource(ActivitySourceName);
+            var activity = activitySource.StartActivity(spanName, ActivityKind.Internal, context);
             activity.Stop();
             return activity;
         }
@@ -164,7 +216,8 @@ namespace OpenTelemetry.Trace.Test
         private Activity CreateNotSampledEndedSpan(string spanName)
         {
             var context = new ActivityContext(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.None);
-            var activity = this.activitySource.StartActivity(spanName, ActivityKind.Internal, context);
+            using ActivitySource activitySource = new ActivitySource(ActivitySourceName);
+            var activity = activitySource.StartActivity(spanName, ActivityKind.Internal, context);
             activity.Stop();
             return activity;
         }